backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
diffusers
transformers
numpy
Pillow
python-dotenv